# spring-server/target/classes/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
  /badges:
    get:
      description: 
        The Badges endpoint returns information about badges
        which are available. The response includes the id, the name, the description, and the image of each badge. This response is a list of badge sorted by the id.
      produces:
        - application/json
      tags:
        - Badges
      responses:
        200:
          description: OK, an array of badges.
          schema:
            type: array
            items:
              $ref: '#/definitions/BadgeOutputDTO'
        500:
          description: If database isn't reachable.
              
    post:
      description: Create a new badge.
      consumes:
        - application/json
      parameters:
        - in: body
          name: badge
          description: Badge object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/BadgeInputDTO'
        - name: token
          in: header
          type: integer
          format: int64
          description: The application for which the badge belong
          required: true
      responses:
        201:
          description: Created, returns the URL to access the new badge.
          schema:
            $ref: '#/definitions/LocationBadge'
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if badge name provided already exists.
        500:
          description: If database isn't reachable.
  /badges/{id}:
    get:
      description: Returns the badge which has the specified id.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the desired badge.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/BadgeOutputDTO'
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
          
    put:
      description: Modify completely the existing badge which has the specified id.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the badge to modify.
          required: true
          type: string
        - name: badge
          in: body
          description: Badge to be updated
          required: true
          schema:
              $ref: '#/definitions/BadgeInputDTO'
          
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if badge name provided already exists.
          

    delete:
      description: Delete the badge which has the specified id.
      parameters:
        - name: id
          in: path
          description: The id of the desired badge.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
...
  /applications:
    get:
      description: 
        The Applications endpoint returns information about an application which are available. The response includes the id, the name, the description. This response is a list of applications sorted by the id.
      produces:
        - application/json
      tags:
        - Applications
      responses:
        200:
          description: OK, an array of application.
          schema:
            type: array
            items:
              $ref: '#/definitions/ApplicationOutputDTO'
        500:
          description: If database isn't reachable.
          
    post:
      description: Create a new application.
      consumes:
        - application/json
      parameters:
        - in: body
          name: application
          description: Application object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/ApplicationInputDTO'
      responses:
        201:
          description: Created, returns the URL to access the new application.
          schema:
            $ref: '#/definitions/LocationApplication'
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if mandatory fields contain more than 80 characters,
                       if application name provided already exists.
        500:
          description: If database isn't reachable.
  /applications/{id}:
    get:
      description: Returns the application which has the specified id.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the desired application.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ApplicationOutputDTO'
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
          
    put:
      description: Modify completely the existing application which has the specified id.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the application to modify.
          required: true
          type: string
        - name: application
          in: body
          description: Application to be updated
          required: true
          schema:
              $ref: '#/definitions/ApplicationInputDTO'
              
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
          schema:
            $ref: '#/definitions/ApplicationInputDTO'
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if application name provided already exists.
          

    delete:
      description: Delete the application which has the specified id.
      parameters:
        - name: id
          in: path
          description: The id of the desired application.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
  /applications/{id}/badges:
    get:
      description: Returns all the badges of the application having the specified id.
      produces:
        - application/json
      tags:
        - Applications
      parameters:
        - name: id
          in: path
          description: The id of the desired application.
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK Array of badges
          schema:
            type: array
            items:
              $ref: '#/definitions/BadgeOutputDTO'
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
...
definitions:
  BadgeOutputDTO:
    type: object
    required: [badge_id, name, description, imageURL]
    properties:
      badge_id:
        type: string
        description: Unique identifier representing a specific badge.
      name:
        type: string
        description: Display name of badge.
      description:
        type: string
        description: Description of badge.
      imageURL:
        type: string
        description: Image URL representing the badge.
  BadgeInputDTO:
    type: object
    required: [name,description,imageURL]
    properties:
      name:
        type: string
        description: Display name of badge.
      description:
        type: string
        description: Description of badge.
      imageURL:
        type: string
        description: The URL of the image          
  LocationBadge:
      type: string
      required: [locationBadge]
      properties:
        locationBadge:
          type: string
          description: Unique URL to access the new object created.
...
  ApplicationOutputDTO:
    type: object
    required: [application_id, name, description]
    properties:
      application_id:
        type: integer
        format: int64
        description: Unique identifier representing a specific application.
      name:
        type: string
        description: Display name of the application.
      description:
        type: string
        description: Description of the application.
  ApplicationInputDTO:
    type: object
    required: [name,description]
    properties:
      name:
        type: string
        description: Display name of the application.
      description:
        type: string
        description: Description of the application.
  LocationApplication:
      type: string
      required: [locationApplication]
      properties:
        locationApplication:
          type: string
          description: Unique URL to access the new object created.
...